Sortt

+++ Sorting.cpp/Bucketsortimproved.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 void bucket(float *arr,int n){
-    int max=arr[0],min=arr[0];
+    float max=arr[0],min=arr[0];
     for(int i=1;i<n;i++){
         if(arr[i]<min){min=arr[i];};
         if(arr[i]>max){max=arr[i];};
